circle

+++ lab5/shapes.h
@@ -19,27 +19,27 @@ public:
 class Rectangle : public Shape
 {
 private:
-	int width, height;
+	float width, height;
 public:
-	Rectangle(Point2D inPoint, int inRgb[], int width, int height) : Shape(inPoint, inRgb), width(width), height(height) {};
+	Rectangle(Point2D inPoint, int inRgb[], float width, float height) : Shape(inPoint, inRgb), width(width), height(height) {};
 	void render(SDL_Renderer* renderer);
 };
 
 class Triangle : public Shape
 {
 private:
-	int base, height;
+	float base, height;
 public:
-	Triangle(Point2D inPoint, int inRgb[], int base, int height) : Shape(inPoint, inRgb), base(base), height(height) {};
+	Triangle(Point2D inPoint, int inRgb[], float base, float height) : Shape(inPoint, inRgb), base(base), height(height) {};
 	void render(SDL_Renderer* renderer);
 };
 
 class Circle : public Shape
 {
 private:
-	int radius;
+	float radius;
 public:
-	Circle(Point2D inPoint, int inRgb[], int radius) : Shape(inPoint, inRgb), radius(radius) {};
+	Circle(Point2D inPoint, int inRgb[], float radius) : Shape(inPoint, inRgb), radius(radius) {};
 	void render(SDL_Renderer* renderer);
 };
